🐛 fix: 상품 별 쿠폰 적용이 아닌 전체 쿠폰 적용
