+++ src/components/features/orderConfirm/orderItem/OrderItem.tsx
@@ -1,16 +1,12 @@
-import { useState } from 'react';
 import { isValidImageUrl } from '../../../../utils/isValidImageUrl';
 import Separator from '../../../common/separator/Separator';
 import { CartItemType } from '../../cart/types';
 import * as S from './OrderItem.styles';
 import defaultImage from '/assets/default_product.png';
-import CouponModal from '../couponModal/CouponModal';
 
 interface OrderItemProps extends CartItemType {}
 
 function OrderItem({ id, quantity, product }: OrderItemProps) {
-  const [isOpen, setIsOpen] = useState(false);
-  const toggleModal = () => setIsOpen((prev) => !prev);
   return (
     <S.Container data-testid={`OrderItem-${id}`}>
       <Separator />
@@ -35,8 +31,6 @@ function OrderItem({ id, quantity, product }: OrderItemProps) {
           <S.Text>{quantity}개</S.Text>
         </S.InfoBox>
       </S.InfoContainer>
-      <S.CouponButton onClick={toggleModal}>쿠폰 적용</S.CouponButton>
-      {isOpen && <CouponModal />}
     </S.Container>
   );
 }
