+++ src/components/features/orderConfirm/orderList/OrderList.styles.ts
@@ -17,3 +17,14 @@ export const AllSelectText = styled.label`
   font-size: 12px;
   font-weight: 500;
 `;
+
+export const CouponButton = styled.button`
+  width: 100%;
+  height: 48px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 15px;
+  font-weight: 700;
+  border: 1px solid #33333340;
+`;
